Export surrealdbNodeEngines from index.js
The surrealdbNodeEngines function was defined and exported in ./lib/embedded.js, but it was not correctly exported in index.js. This caused issues when trying to use surrealdbNodeEngines from the @surrealdb/node package. By exporting it correctly, we ensure that the function is available for import and use in other parts of the application.